mock-cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "databao-mock-cli"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2",
    "pyyaml>=6.0",
    "dbt-core>=1.0",
    "questionary>=2.0",
]

[project.scripts]
databao = "databao_mock.__main__:cli"

[tool.hatch.build.targets.wheel]
packages = ["databao_mock"]
